+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.7.0 (not released yet)
 
+* API change: Renamed `displayName` to `label`.
 * API change: The `ActionArgMeta` class renamed to `ArgMeta`.
 * API change: A provided argument specification in now placed in the `ArgProvided` class, not directly in the `ArgMeta` as before.
 * API change: Removed `ArgValue.valueSetDisplayNames` because of a new support for a labeled value set.